+++ src/pages/community/article/Article.scss
@@ -51,6 +51,10 @@
               cursor: pointer;
             }
           }
+
+          .none {
+            display: none;
+          }
         }
 
         .titleInner {
