+++ src/pages/community/article/CommentInput.scss
@@ -1,0 +1,87 @@
+@import '../../../styles/mixin.scss';
+@import '../../../styles/variables.scss';
+
+.commentPage {
+  padding: 30px;
+
+  .userInfo {
+    @include flex(normal, center);
+
+    .profileImg {
+      width: 50px;
+      height: 50px;
+      margin-right: 10px;
+      border-radius: 50%;
+    }
+
+    .tier {
+      margin-right: 5px;
+    }
+
+    .userName {
+      font-size: 18px;
+    }
+  }
+
+  .comment {
+    margin-top: 20px;
+    border: 1px solid $white;
+
+    .commentInput {
+      width: 100%;
+      height: 100px;
+      margin: 0 auto;
+      padding: 15px;
+      border: none;
+      resize: none;
+      font-size: 18px;
+      font-family: 'Noto Sans KR', sans-serif;
+      font-weight: 400;
+
+      &:focus {
+        outline: none;
+      }
+
+      &::placeholder {
+        font-family: 'Noto Sans KR', sans-serif;
+      }
+    }
+
+    .enroll {
+      text-align: right;
+
+      .enrollBtn {
+        margin: 10px;
+        width: 70px;
+        padding: 8px;
+        border: none;
+        border-radius: 20px;
+        background-color: #eee;
+        color: gray;
+        font-weight: bold;
+      }
+
+      .active {
+        background-color: $pointColor;
+        color: white;
+        cursor: pointer;
+      }
+    }
+  }
+
+  .loginMsg {
+    @include flex(normal, center);
+    padding: 20px 30px;
+    border-radius: 7px;
+    background-color: $commonBg;
+
+    .icon {
+      font-size: 20px;
+    }
+
+    .goToLogin {
+      margin: 7px;
+      font-weight: bold;
+    }
+  }
+}
