+++ src/pages/community/article/CommentList.tsx
@@ -1,10 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import axios from 'axios';
-import { FaThumbsUp, FaRegThumbsUp, FaRegComment } from 'react-icons/fa';
-import { FiCornerDownRight } from 'react-icons/fi';
-import ReComment from './ReComment';
 import './CommentList.scss';
+import Comment from './Comment';
 
 type Comment = {
   id: number;
@@ -23,6 +21,13 @@ type Recomment = {
   content: string;
 };
 
+export type CommentProps = {
+  comment: Comment;
+  idx: number;
+  postId: number;
+  loadComment: void;
+};
+
 export type ReCommentProps = {
   data: { id: number; name: string; tier: string; content: string };
 };
@@ -31,7 +36,6 @@ const CommentList = () => {
   const [commentList, setCommentList] = useState<Comment[]>([]);
   const [copyCommentList, setCopyCommentList] = useState<Comment[]>([]);
   const [currentTab, setCurrentTab] = useState<number>(0);
-  const [reComOpen, setReComOpen] = useState<null | number>(null);
   const postId = useParams<string>();
   const token = localStorage.getItem('token');
 
@@ -55,35 +59,9 @@ const CommentList = () => {
     });
   };
 
-  // 댓글 삭제
-  const deleteComment = (idx: number) => {
-    alert('댓글을 삭제하시겠습니까?');
-    axios
-      .delete(`/community/comments/${idx}`)
-      .then(res => loadComment())
-      .catch(err => console.log(err));
-  };
-
-  // 대댓글 토글
-  const toggleReCom = (idx: number) => {
-    setReComOpen(prev => (prev === idx ? null : idx));
-  };
-
   useEffect(() => {
     loadComment();
   }, []);
-
-  let numArr = commentList.map(x => x.likesNum);
-  const [isClicked, setIsClicked] = useState<number | null>(null);
-  const [change, setChange] = useState<number[]>(numArr);
-
-  const clickIcon = (idx: number) => {
-    // 아이콘 변경
-    setIsClicked(prev => (prev === idx ? null : idx));
-    let copy = [...numArr];
-    copy[idx]++;
-    setChange(copy);
-  };
 
   return (
     <>
@@ -102,62 +80,14 @@ const CommentList = () => {
       </nav>
       <div className="commentList">
         {commentList.map((comment, idx) => {
-          const { id, name, img, content, createdAt, likesNum, reComment } =
-            comment;
           return (
-            <div key={id}>
-              <div className="comment">
-                <section className="userInfo">
-                  <img className="profileImg" src={img} alt="profile img" />
-                  <ul className="infoContent">
-                    <li className="tier">Tier</li>
-                    <li className="userName">{name}</li>
-                    <li className="time">{createdAt}</li>
-                    <li
-                      className="deleteBtn"
-                      onClick={() => deleteComment(idx)}
-                    >
-                      삭제
-                    </li>
-                  </ul>
-                </section>
-                <div className="content">{content}</div>
-              </div>
-              <section className="reComHeader">
-                <div className="thumbsUpWrap">
-                  {isClicked === idx ? (
-                    <FaThumbsUp
-                      className="thumbsUp"
-                      onClick={() => clickIcon(idx)}
-                    />
-                  ) : (
-                    <FaRegThumbsUp
-                      className="thumbsUp"
-                      onClick={() => clickIcon(idx)}
-                    />
-                  )}
-                </div>
-                <span>{change[idx]}</span>
-                <div className="reComBtn" onClick={() => toggleReCom(idx)}>
-                  <FaRegComment />
-                  <span>댓글 달기</span>
-                </div>
-              </section>
-              <div className={reComOpen === idx ? '' : 'hidden'}>
-                <div className={token ? 'writeReCom' : 'hidden'}>
-                  <FiCornerDownRight className="writeReComIcon" />
-                  <form className="reComForm">
-                    <textarea className="reCom" placeholder="댓글 남기기" />
-                    <div className="enroll">
-                      <button className="enrollBtn">등록</button>
-                    </div>
-                  </form>
-                </div>
-                {reComment.map(data => {
-                  return <ReComment key={data.id} data={data} />;
-                })}
-              </div>
-            </div>
+            <Comment
+              key={idx}
+              idx={idx}
+              postId={postId}
+              comment={comment}
+              loadComment={loadComment}
+            />
           );
         })}
       </div>
